refactor: move example plugin to playground
