+++ packages/api-reference/playground/vue/src/x-custom-extension-plugin/components/XCustomExtensionComponent.vue
